+++ tests/test_benchmark_simple_scenarios.py
@@ -7,15 +7,18 @@ What should differ between the different functions is the input file
 import argparse
 import os
 import shutil
+import json
 
 import mock
 import pandas as pd
 import pytest
 
 from mvs_eland_tool import main
+from src.B0_data_input_json import convert_special_types
 from .constants import (
     EXECUTE_TESTS_ON,
     TESTS_ON_MASTER,
+    TESTS_ON_DEV,
     TEST_REPO_PATH,
     CSV_EXT,
 )
@@ -82,6 +85,32 @@ class TestACElectricityBus:
         df_busses_flow = df_busses_flow.set_index("Unnamed: 0")
 
         # TODO make sure the battery is not used in an assert here
+
+    @pytest.mark.skipif(
+        EXECUTE_TESTS_ON not in (TESTS_ON_MASTER),
+        reason="Benchmark test deactivated, set env variable "
+        "EXECUTE_TESTS_ON to 'master' to run this test",
+    )
+    @mock.patch("argparse.ArgumentParser.parse_args", return_value=argparse.Namespace())
+    def test_benchmark_ABE_grid_pv_bat(self, margs):
+        # define the two cases needed for comparison (grid + PV) and (grid + PV + battery)
+        use_case = ["AB", "ABE_grid_PV_battery"]
+        # define an empty dictionary for excess electricity
+        excess = {}
+        for case in use_case:
+            main(
+                path_input_folder=os.path.join(TEST_INPUT_PATH, case),
+                input_type=CSV_EXT,
+                path_output_folder=os.path.join(TEST_OUTPUT_PATH, case),
+            )
+            busses_flow = pd.read_excel(
+                os.path.join(TEST_OUTPUT_PATH, case, "timeseries_all_busses.xlsx"),
+                sheet_name="Electricity bus",
+            )
+            # compute the sum of the excess electricity for all timesteps
+            excess[case] = sum(busses_flow["Electricity excess_sink"])
+        # compare the total excess electricity between the two cases
+        assert excess["AB"] < excess["ABE_grid_PV_battery"]
 
     def teardown_method(self):
         if os.path.exists(TEST_OUTPUT_PATH):